Fix sensor BLE timestamp packing (us->ms), divider type consistency, doc alignment
- Convert sensor stream timestamps from raw uint64 microseconds to uint32
  milliseconds before packing into BLE payload, matching documented protocol
  (accel_ble, baro_ble, gyro_ble, hum_ble, mag_ble)
- Fix ACCEL/GYRO/MAG BLE decimation divider locals from uint8 to uint16 to
  avoid truncation of large divider values (custom_app.c)
- Update ble.md: add timestamp epoch/source semantics section, correct baro
  max-length (10->12), baro pressure units (Pa->Pa*100), hum/mag temperature
  units (0.01C->0.1C)

+++ STM32_WPAN/App/custom_app.c
@@ -1019,7 +1019,7 @@ void Custom_BARO_Update(const FS_Baro_Data_t *current)
 void Custom_ACCEL_Update(const FS_IMU_Data_t *current)
 {
   /* Decimation: only send every Nth sample */
-  uint8_t divider = ACCEL_BLE_GetDivider();
+  uint16_t divider = ACCEL_BLE_GetDivider();
   if (++accel_sample_counter < divider)
   {
     return;
@@ -1038,7 +1038,7 @@ void Custom_ACCEL_Update(const FS_IMU_Data_t *current)
 void Custom_GYRO_Update(const FS_IMU_Data_t *current)
 {
   /* Decimation: only send every Nth sample */
-  uint8_t divider = GYRO_BLE_GetDivider();
+  uint16_t divider = GYRO_BLE_GetDivider();
   if (++gyro_sample_counter < divider)
   {
     return;
@@ -1057,7 +1057,7 @@ void Custom_GYRO_Update(const FS_IMU_Data_t *current)
 void Custom_MAG_Update(const FS_Mag_Data_t *current)
 {
   /* Decimation: only send every Nth sample */
-  uint8_t divider = MAG_BLE_GetDivider();
+  uint16_t divider = MAG_BLE_GetDivider();
   if (++mag_sample_counter < divider)
   {
     return;
